internal/jem: add support code for monitor work.
Add ErrAPIConnection as an error cause from OpenAPI
so that we can distinguish between a JEM database error and
an error when trying to connect to the Juju controller.

Add a Clone method so that a long-lived request
can create a JEM instance from another one, suitable for
using asynchronously.

Add AcquireMonitorLease to support acquiring the
lease when monitoring a controller.

Add SetControllerStats to support setting statistics
associated with a controller.

Add SetModelLife to support setting the life status
associated with a model.

Also, in internal/apitest, add ProxiedPool method to make it easier
to test JEM connections when the network breaks and expose a
Pool to tests.