internal/monitor: fix panic in monitor when controller is removed
We're not allowed to return ErrDying from a tomb goroutine
when the tomb isn't already dying.

Fix this by returning ErrDying only when we are already dying,
and returning another error otherwise. We check the error
in the allMonitor, which means that we'll print a more meaniningful
error when a controller monitor dies.