internal/v1: expose life and availability in API
We also make it so that you can't delete a controller when
it's alive unless you pass the force flag.

And we don't run the monitor by default so that it
can't interfere with tests. We choose this default because
we only have the production config in one place
but tests create ServerParams all over the place.