internal/monitor, internal/apiconn: be resilient to broken controller connections
Controller TCP connections can be broken.
This change makes the API connection cache refuse to return
connections that are known to be broken,
and the monitor now evicts API connections that are known bad.

Ideally the juju API client would mark a connection as broken
as soon as it gets a bad reply from the server (that would make
the Evict call in monitor/controller.go unnecessary), but that's something
that can happen later.