internal/jem: change monitor lease representation
We change the monitor lease fields in the mongo controller
document so that they're omitted when the lease isn't taken.
This means that the database is backwardly compatible
when upgraded, and also means that we don't store
zero time.Time values in the database (they don't look
very nice).

JEM instances that have already been deployed with
the monitoring enabled will need some manual database
tweaks to remove the zero fields.