internal/monitoring: rename to internal/servermon
Given that we have both "monitpr" and "monitoring" packages,
the distinction between the two isn't so clear, so we rename
the newer of the two.

We also change the Request methods so they're hopefully a little clearer
and make the monitor debug messages print out the full contents
of the deltas.