internal/jem: new scheme for session sharing
The current scheme has the issue that the code has no idea when
a session has become unusable, so a bad session will continue
to be used for some time after a mongo connection has been
severed.

This PR proposes a new scheme where we check the error after
any database operation; if there's an error that we don't recognise,
then it's probably because of a database error, so we avoid reusing
that session, making a new copy instead.

We maintain a number of current sessions that we use in round-robin
fashion so not all clients will be using the same session.

In a subsequent PR, we will change the internal/auth package to
use the same scheme, possibly factoring out some of the existing
logic into something more reusable.