cmd/jemd: use zap for logging
We want to have structured, context-sensitive logging
and the zap package provides that in a decent way.

We set up loggo so that it will log to a zap logger
and create a new zapctx package that will log
messages to loggers found in Context values.

We add a new logging-level value to the configuration
file and deprecate the logging-config flag (printing
a warning if it's set).

An example logging line, as logged through loggo:

    {"level":"info","ts":1479908675.7637515,"msg":"dialing \"wss://10.0.8.99:17070/api\"","module":"juju.api","file":"/home/rog/src/go/src/github.com/juju/juju/api/apiclient.go","line":530}