internal/mgosession: throw away all sessions on error
There is a problem with the current implementation that if the mongo
server restarts, all the sessions become invalid at the same point,
but we will only find out about each one in turn.

Therefore we change the mgosession package so that it throws
away all the sessions when an error is encountered,
which enables some simplification of the code.

We also start a pinger which resets the pool if it encounters
an error, which should limit the maximum time a bad mongo
session causes problems, even when we don't reset the
pool immediately because we diagnose a bad error.

This may fix some operational problems we've been seeing.