internal/monitor: fix deadlock when mongo connection dies
We were not waiting for the controller monitors to
exit when returning an error due to AllControllers failing.

We also make sure that the controller monitors exit
when the allMonitor is exiting by killing them explicitly.
That should not have been a real problem because in
practice the controller monitors will also die when
the mongo connection goes away, but it's better to
be complete (and it means that the controller monitor
can be restarted immediately even when there's an
API open call that's taking a long time to complete.