internal/jem: make CreateModel more resilient to juju controller API failure
By inspection, the model creation code is vulnerable to an
issue where the model isn't removed if we fail to grant the
required permissions to it, and it isn't updated to the
correct details either.

We change CreateModel to remove the model on any error
It is still possible that the model remains in the controller
after this - for example if we failed to grant correct access to it.
If that happens (hopefully very rare), we return an error
that indicates the problem. In the future, we might create
a worker to resolve that issue automatically.