internal/monitor: remove models even when we don't get a watch update
If a controller's watcher is broken or the jimm instance is restarted
after a model has been deleted but before we've received the
life change, we can end up with an entry in the jimm database
for which there is no corresponding controller model.

Fix this by enumerating all the local models before we start watching,
so that they'll be deleted if we don't see a model update in the
first watcher delta (the delta which has all information on controller).