cmd/jemd: improve application lifecycle.
Update the main server application to allow it to be shutdown
gracefully. The server can now be stopped gracefully using the INT or
TERM signals which will gracefully shutdown the service.

This update also removed the joining of per-request context to the root
context. The root context will now be closed on a signal, but we want to
ensure that individual requests are stopped gracefully rather than the
moment a signal is received.