internal/cloudcred: credential attributes table
Previously all the known supported providers were imported into JIMM.
These come with a lot of dependencies. The only reason for having these
is to be able to determing if credential attributes should be redacted
or not.

By precomputing this, jemd no longer needs to import all the providers.
This reduces the jemd binary size from ~100M to ~67M.