internal/mongodoc: change application-offer access storage
Change the way that application-offers store user access to be a map.
This provides a simpler mechanism for updating and checking access at
the cost of having to encode usernames such that they can be safely
stored in mongodb.

This change is safe to make at this time as application-offers are not
yet available in production.