internal/jujuapi: JIMM facade AddController method
Create an AddController method on the JIMM facade that allows a JAAS
controller-admin to add new juju controllers to the JAAS system. This
is equivalent to the PUT /v2/controller/:User/:Name endpoint. The
intention is to allow controller management to be more easily integrated
with the dashboard and to remove the /v2 HTTP API.