internal/db: add initial database access methods
Adds the initial database access methods. Currently only database
migrations are supported. This includes code for testing all the
database systems we are planning to support.

This also adds a new nascent interface for errors that tries to address
errors on a system-wide basis. This interface is inspired by
https://commandcenter.blogspot.com/2017/12/error-handling-in-upspin.html
and https://middlemost.com/failure-is-your-domain/.