internal/rpc: add cancellable juju RPC client
Add a new Juju RPC client implementation that uses contexts such that
requests can be cancelled on the client end and return early. The
protocol itself has no facility to cancel requests so the procedure
will still run to completion on the server.

This will enable a much simpler monitor process to be implemented.

As there is no existing reference for the juju RPC protocol this has
also been documented.