all: update linked juju to 2.9-rc7
Update the juju version to 2.9-rc7. This also changes the mgo dependency
to be github.com/juju/mgo/v2 to match juju, that resulted in a lot of
module churn.

For some reason this update slows tests down considerably so the timeout
has been increased until that can be rectified.