config: enable multiple admin users
The original conception for JIMM admin users was that it could be a
single group in an upstream identity provider. With the increase in
identity providers, and more diverse deployment scenarios this is now
less practical. JIMM now supports providing a list of users, or groups,
who require admin access.