internal/dbmodel: remove machines and units
Stop storing machine an unit information in the database. Instead just
store the few statistics needed to server model summaries. This requires
changing the way ModelInfo is implemented, this will now collect the
ModelInfo from the hosting controller rather than from the database. This
is safe to do as ModelInfo is no longer used in juju list-models.