CSS-737 Fix application-offer permissions logic
The application-offer permissions logic was inconsistent and buggy:
1. A controller superuser or model admin could (theoretically) grant/revoke access to an offer, but they were not able to find the offer.
2. A bug in the model fetching logic meant the model admins couldn't actually perform the above.
3. Unit tests were not capturing the bugs above.

This PR fixes all issues, such that now either a controller superuser, a user that has admin rights on the model, or admin access to the offer can find the offer and all its user, and can grant or revoke access from said offer, or delete the offer completely.

I also edited the tests so that they capture the above issues as well.