Fix for the AddModel cleanup.
In case our request to create a model on the controller times out,
we try to delete the model from JIMM's DB. Since we were using
the expired context, the call to delete the model would fail.

Also increased request timeout to 30s as we've seen slower
responses from Juju controllers.