feat(auth service device flow): introduces an auth service capable of performing oauth2 device flow
To isolate the oidc and oauth2 logic, a new service has been created with a single flow currently
(device) and a means to verify the id token, and additionally extract the email. It is currently
uncertain whether we should ensure the email is bverificed first.   The test is very much an
all-in-one test as it semantically makes sense to do the entire flow in a single test for the device
flow.