Refactor JWKS cache (#1148)
Refactor the JWKS cache away from the jwk.Cache object which caches an HTTP response and causes JIMM to make an HTTP call to itself. Instead, we use a cache with a TTL and fetch the JWK set from Vault periodically (currently set to 1h) with a frequency much higher than the cache refresh interval.