feat(logindevice): implements LoginDevice
Login device now stores the oauth response on the controllerroot such that it can be shared with the
second step of the flow, GetDeviceAccessToken upon a successful login from a user. It is expected
that the same websocket be used for both facade calls. This prevents the need to associate a
LoginDevice call to a GetDeviceAccessToken. In addition to this, the library actually expects the
oauth response from device to be sent to deviceaccesstoken.

6714