feat(oauth login browser): implements /auth/login (#1166)
* feat(oauth login browser): implements /auth/login

As discussed on call, we integration test only the handler and avoid mocks to see the behaviour is
as expected. This is also true when we come to implements the callback logic and are required to
start the flow from scratch. The current test in auth_handler_test will be updated to cover the
entire flow when implementing /callback. As for the state todo, I need to see how to track the state
between requests and the TODO will be completed in the /callback PR.

6646

* pr comments

* pr comments