Css 6646/callback endpoint (#1170)
* feat(oauth login browser): implements /auth/login

As discussed on call, we integration test only the handler and avoid mocks to see the behaviour is
as expected. This is also true when we come to implements the callback logic and are required to
start the flow from scratch. The current test in auth_handler_test will be updated to cover the
entire flow when implementing /callback. As for the state todo, I need to see how to track the state
between requests and the TODO will be completed in the /callback PR.

6646

* pr comments

* pr comments

* feat(browser flow for dashboard): implements browser flow (without sessions)

This PR includes a small refactor to the admin device flow, such that it can share the same identity
update logic within the browser flow.

* feat(validation in auth handler): validates params given are correct and auth svc not nill

* Fix tests

* pr comments

* additional failure tests

* test fix

* Update refresh token