Ensure the latest tag is not replaced with a lightweight tag
Due to the issues mentioned in the comments, when using the Github checkout action, when checking out via tag, the desired tag is converted to a lightweight tag even if it was originally an annotated tag. This causes git describe to break which also causes tools like snapcraft's versioning to break.