fix cmd tests
Juju changed their CLI login in https://github.com/juju/juju/pull/17667 to fix a bug so now CLI commands check that the user in the accounts.yaml file you were trying to login as, matches the user you ended up logging in as.
In JIMM's tests, we have an in-memory accounts struct that was empty so this was causing an issue.