Remove db from jujuapi (#1297)
* Avoid use of DB outside of jimm

* test fixes

Replace hardcoded controller UUID in tests with constant

* Export ControllerService

Drive-by warning fix

* minor refactoring

* avoid returning a params type from jimm layer

Rather than returning an apiParams type from the jimm layer in `ControllerInfo`, return a dbmodel. This also means it doesn't make sense to return the JIMM controller info since that is not a entity stored in the database. The construction of the JIMM controller info is kept directly in `jujuapi`.

* update godocs