feat: introduce a model manager domain service
The model manager package in internal/jimm is the first for a Juju related domain that moves logic away from the large jimm struct into a dedicated package. This PR shows that the model manager relies on bits from what will eventually be the cloud and controller managers and also requires a way to dial Juju controllers.

This change is incomplete because I'd like the team to get a feel for whether they enjoy this approach of a smaller package with more dependencies OR whether they'd prefer a larger (i.e. single) Juju package that has less dependencies.