chore: create juju business logic package
This PR moves all the business logic related to Juju resources into a package called juju. The naming is a bit poor imo. The first thought that comes to mind is that we should rename the `jimm` package to something like `app` or `business` to indicate that it holds the business logic.