fix: add limit listener tests (#1550)
* fix: fix limit listener with timeout

The limit listener with timeout was not working properly when it reached saturation, at which point new connections would be closed even if old connections were closed.
Adds a test for this and copies the existing set of tests from the original implementation.

* chore: tweak limit listener accept()

Tweak the internal accept() method on the limit listener to return two bools so that we can check without ambiguity whether the listener timed out or was closed.

* chore: update godoc