chore: improve doc publishing and jaas tool detection (#1594)
When using the JAAS tool standalone, we should directly reference the binary, using the file at /snap/bin/jaas references the snapcraft runner which intentionally doesn't work as the jaas snap is intended to be run via the Juju CLI as a plugin.
Changes the documentation generation to put the docs into the correct file.
Update some commands where tabs and spaces were mixed. Spaces seem to be the correct and more widely used option.