feat(qa-script): adds qa scripts for mac (arm) developers
There are two scripts. qa-lxd-mac and qa-lxd-mac-forward.sh. The first spins up a multipass VM, runs
jimm within compose inside and adds a controller. The second port-forwards JIMM and keycloak, and is
expected to be run AFTER  qa-lxd-mac, such that you can connect and QA. The VM has a volume mount
and changes will be reflected and updated from the hosts jimm repo.