chore: remove special case for service accounts
The special case service accounts tag was used to manage a service account from a user's account. This was done so that a user could create cloud-credentials for a service account. Now a service account can be fully managed by Terraform so there is no need for all this special casing which probably shouldn't have been done in the first place in any case.