chore: partially move to the Juju api client
Uses the cloud methods to illustrate what a move to the juju api client might look like.
Some challenges included the need to change the types we use internally from jujurpc types to more concrete definitions as used by the client.
This also required a way to convert from the rpc type that we receive from the client to the concrete type.
The converter function exists in Juju in 2 places (the client and apiserver) but both are unexported and copy-paste of each other so I've also copied it into JIMM for now.
Also removed the redial logic as I'm confident it's not necessary.