chore: remove context from cloud calls
It was discussed that it doesn't make sense to use a context for websocket messages where there is no notion of cancelling a single message in flight.
Instead, the context of a message is tied to the life/context of the overall websocket.
So we remove the context from the updated API methods (with the rest to come in future changes).