feat: prevent controller deletions during migrations
A controller could be deleted mid migration, this would leave us in a poor spot. This prevents the
controller being unregistered if a migration is active. It is to be plugged into the prepare call
(to set the migration active) and to be unset on the final Active call during a migration.