chore: partially move to the Juju api client (#1611)
* chore: partially move to the Juju api client

Uses the cloud methods to illustrate what a move to the juju api client might look like.
Some challenges included the need to change the types we use internally from jujurpc types to more concrete definitions as used by the client.
This also required a way to convert from the rpc type that we receive from the client to the concrete type.
The converter function exists in Juju in 2 places (the client and apiserver) but both are unexported and copy-paste of each other so I've also copied it into JIMM for now.
Also removed the redial logic as I'm confident it's not necessary.

* chore: remove context from cloud calls

It was discussed that it doesn't make sense to use a context for websocket messages where there is no notion of cancelling a single message in flight.
Instead, the context of a message is tied to the life/context of the overall websocket.
So we remove the context from the updated API methods (with the rest to come in future changes).

* chore: remove cache dialer

* chore: test fixes

* chore: remove unused WithContext method