feat: redirect agents to backing controller
This change is in support of model migration. During model migration, agents connect to the new controller to validate they will work after migration has completed.
JIMM should not be the address for Juju agents, they should point to the backing controller but the migration flow needs to go through JIMM for it to modify things like the model owner.
To handle this, when an agent is trying to login to JIMM we return a redirect to the backing controller.