feat: `qa-lxd-multipass` scripts (#1638)
* feat(qa-script): adds qa scripts for mac (arm) developers

There are two scripts. qa-lxd-mac and qa-lxd-mac-forward.sh. The first spins up a multipass VM, runs
jimm within compose inside and adds a controller. The second port-forwards JIMM and keycloak, and is
expected to be run AFTER  qa-lxd-mac, such that you can connect and QA. The VM has a volume mount
and changes will be reflected and updated from the hosts jimm repo.

* chore: new line

* fix: don't crash out on missing path

* fix: accept unknown host keys when testing SSH tunnels

* fix: use `VM_NAME` when generating multipass mount name

* fix: install go into multipass vm

* fix: proxy ports in single command

* fix: correct logic for prepending URL scheme

* fix: optionally provide VM name

* fix: keep SSH tunnels open while logging in to juju CLI

* fix: capitals to stop docker complaining

* fix: detect OS with `uname` if flag not provided

* fix: cross compile `jaas` for linux

* fix: rename `qa-lxd-mac` to `qa-lxd-multipass` since it's OS agnostic

* fix: use heredoc to execute multiple commands in multipass

* fix: add `qa-multipass` and `qa-multipass-forward` make targets

---------

Co-authored-by: ale8k <alexjkilroy@gmail.com>