fix: add info field to wellknown handler tests
The JWKS endpoints return a JSON object that is just mapping our internal error type.
We shouldn't do this, there is no specification that requires our JWKS endpoint to return a JSON formatted error and we especially shouldn't just marshal our internal error type to JSON.
This will need to be addressed separately in a follow-up.