feat: redirect agents to backing controller (#1643)
* feat: redirect agents to backing controller

This change is in support of model migration. During model migration, agents connect to the new controller to validate they will work after migration has completed.
JIMM should not be the address for Juju agents, they should point to the backing controller but the migration flow needs to go through JIMM for it to modify things like the model owner.
To handle this, when an agent is trying to login to JIMM we return a redirect to the backing controller.

* chore: rename servers to addresses

* fix: add info field to wellknown handler tests

The JWKS endpoints return a JSON object that is just mapping our internal error type.
We shouldn't do this, there is no specification that requires our JWKS endpoint to return a JSON formatted error and we especially shouldn't just marshal our internal error type to JSON.
This will need to be addressed separately in a follow-up.

* chore: fix rpcproxy tests