chore: add offers and permissions during migration import (#1647)
* chore: add offers and permissions during migration import

* chore: remove modelMigrationImport struct

* chore: explicitly use controller tag

When creating a model, the model's Controller ID needs to be present but our CreateModel function doesn't fetch the controller's details and place this on the model struct, as opposed to GetModel which does.
This means after model creation, if we want the controller tag, we need to either use the controller details that we are certain we have or run GetModel.
Here I've opted to just pass the controller details that we already have to `addModelAndOfferPermissions`.