chore: allow PrepareModelMigration to be called multiple times
Migrating a model to JAAS requires a client to call PrepareModelMigration first.
This change makes it possible to call PrepareModelMigration multiple times and allows the initial user mapping or backing controller to change. JIMM will ensure that if the model migration has progressed up to a certain point that the PrepareModelMigration will return an error instead to ensure the data is consistent.