chore: allow PrepareModelMigration to be called multiple times (#1649)
* chore: allow PrepareModelMigration to be called multiple times

Migrating a model to JAAS requires a client to call PrepareModelMigration first.
This change makes it possible to call PrepareModelMigration multiple times and allows the initial user mapping or backing controller to change. JIMM will ensure that if the model migration has progressed up to a certain point that the PrepareModelMigration will return an error instead to ensure the data is consistent.

* chore: simplify locking clause

* chore: prefetch controller

* chore: check error matches string

* chore: run db method with nested transaction

* chore: use the tx in a transaction

It's easy to accidentally use j.Database when inside a transaction closure, but this means your DB query may use a different connection to the database and not the one where you have started a transaction.
This change fixes that in a few places.