test(name with uuid): add uuid to test's db name and openfga db name
Before I wasn't able to stress test TestGetBootstrapStatusAndLogs because during SetupTestOFGAClient
in RemoveStore it does "DELETE FROM authorization_model WHERE store = (SELECT id FROM store WHERE
name = %s"). This returns an error if the subquery returns multiple entries. And this can happen if
we stress test a test because it is setting up the store multiple times with the same name.