feat: replace SetModelMigrationMode with GetModelForUpdate
The SetModelMigrationMode method was placing logic in the DB layer and after some comments on the first PR I realised that i was introducing inconsistent behavior where some DB logic was captured within the method and other logic was captured on the dbmodel struct.
This change introduces a GetModelForUpdate method that will return the model and lock the row for updates, allowing the caller to handle the business logic while adding some helper methods on the dbmodel.Model struct.