feat: handle migrations from model proxy
In the model proxy it we will return an error when a model has been migrated to inform the client to try again later.
We opt for this approach rather than completing the model migration because the model proxy starts to become over-complicated when it is doing more than just proxy logic unless absolutely necessary.